+++ app/src/main/java/com/github/damontecres/wholphin/ui/detail/movie/MovieDetails.kt
@@ -22,6 +22,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.Dp
@@ -483,6 +484,7 @@ fun MovieDetailsContent(
                         items = similar,
                         showLoadMore = state.similarHasMore,
                         onClickLoadMore = onLoadMoreSimilar,
+                        onLoadMoreFocus = { position = SIMILAR_ROW },
                         onClickItem = { index, item ->
                             position = SIMILAR_ROW
                             onClickItem.invoke(index, item)
@@ -505,7 +507,12 @@ fun MovieDetailsContent(
                         modifier =
                             Modifier
                                 .fillMaxWidth()
-                                .focusRequester(focusRequesters[SIMILAR_ROW]),
+                                .focusRequester(focusRequesters[SIMILAR_ROW])
+                                .onFocusChanged {
+                                    if (it.hasFocus) {
+                                        position = SIMILAR_ROW
+                                    }
+                                },
                     )
                 }
             }